relocate confirm dialog
