+++ src/frontend/src/components/ConfirmDialog/style.ts
